Criando Repository e Service para Catgoria

+++ src/main/java/com/candidowagner/coursespring/domain/Categoria.java
@@ -14,23 +14,23 @@ public class Categoria implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 	
 	public Categoria() {
 	}
 
-	public Categoria(Long id, String nome) {
+	public Categoria(Integer id, String nome) {
 		super();
 		this.id = id;
 		this.nome = nome;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -67,4 +67,6 @@ public class Categoria implements Serializable {
 		return true;
 	}
 	
-}
+	
+
+	}
